implement sync for more than 1 worker
Client sends EOT[n] (end of transmission) msg when it's finished sending the files. [n] denotes the number of workers available to fetch the raw data sent.
Each worker in the receive end will receive an EOT[n] message. If [n] == 1, then this means the current worker is also the last one to receive the msg, hence it's the one responsible to retransmit the EOT msg to the subordinates, i.e. the nodes which receive info from the raw file receivers.
If [n] is not 1, then the worker forms a new msg EOT[n-1] and sends it to the client. The client will fwd this msg to the raw file receivers' queue. This way, all receivers will receive the EOT msg.